added info to ignore file
